Subindo métodos de cadastrar vendedores e buscar

+++ backend/src/main/java/com/wd/projeto01/entities/Venda.java
@@ -16,7 +16,7 @@ public class Venda {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private Integer visitas;
 	private Integer qtdVendas;
 	private Double valor;
@@ -28,7 +28,7 @@ public class Venda {
 	
 	public Venda() {}
 
-	public Venda(Long id, Integer visitas, Integer qtdVendas, Double valor, LocalDate data) {
+	public Venda(Integer id, Integer visitas, Integer qtdVendas, Double valor, LocalDate data) {
 		super();
 		this.id = id;
 		this.visitas = visitas;
@@ -37,11 +37,11 @@ public class Venda {
 		this.data = data;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
