+++ backend/src/main/java/com/wd/projeto01/repositories/VendaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wd.projeto01.entities.Venda;
 
-public interface VendaRepository extends JpaRepository<Venda, Long>{
+public interface VendaRepository extends JpaRepository<Venda, Integer>{
 
 }
